Rename get() to _self(), add _type(), new utils for type checks
Major rename of internal get() method to _self() and added a _type()
method.  This allows for less conflicts and easy identification of
object type.  Added new type check util methods and updated all queries
and filters to use them.  Renamed get() method on the Request object to
search to indicate it is executing a search using the _search endpoint.
 This is useful because we can add other methods to the same Request
object for other endpoints.  For example, we can add count and explain
endpoints re-using the same object.